Move over several helper scripts
