+++ core/scripts/WhitelistSponsor.js
