+++ core/scripts/WhitelistToken.js
