+++ core/test/scripts/ManualPublishPriceFeed.js
